Rename column in stack sample, move to column samples group
